+++ src/main/java/beyond/momentours/comment/command/domain/vo/request/RequestCreateCommentVO.java
@@ -16,9 +16,6 @@ public class RequestCreateCommentVO {
     @JsonProperty("comment_type")
     private CommentType commentType;
 
-    @JsonProperty("ques_id")
-    private Long quesId;
-
-    @JsonProperty("couple_log_id")
-    private Long coupleLogId;
+    @JsonProperty("target_id")
+    private Long targetId;
 }
