+++ src/main/java/beyond/momentours/comment/command/domain/vo/response/ResponseCreateCommentVO.java
@@ -17,9 +17,6 @@ public class ResponseCreateCommentVO {
     @JsonProperty("comment_type")
     private CommentType commentType;
 
-    @JsonProperty("ques_id")
-    private Long quesId;
-
-    @JsonProperty("couple_log_id")
-    private Long coupleLogId;
+    @JsonProperty("target_id")
+    private Long targetId;
 }
