+++ src/main/java/beyond/momentours/date_course/command/application/dto/DateCourseDTO.java
@@ -17,7 +17,6 @@ public class DateCourseDTO {
     private Long courseId;
     private String courseTitle;
     private CourseType courseType;
-    private String courseMemo;
     private Boolean courseDisclosure;
     private Long courseLike;
     private Long courseView;
