+++ src/main/java/beyond/momentours/date_course/command/application/mapper/DateCourseConverter.java
@@ -19,7 +19,6 @@ public class DateCourseConverter {
         return DateCourseDTO.builder()
                 .courseTitle(request.getCourseTitle())
                 .courseType(request.getCourseType())
-                .courseMemo(request.getCourseMemo())
                 .courseDisclosure(request.getCourseDisclosure())
                 .courseStartDate(request.getCourseStartDate())
                 .courseEndDate(request.getCourseEndDate())
@@ -31,7 +30,6 @@ public class DateCourseConverter {
                 .courseId(saveDateCourseDTO.getCourseId())
                 .courseTitle(saveDateCourseDTO.getCourseTitle())
                 .courseType(saveDateCourseDTO.getCourseType())
-                .courseMemo(saveDateCourseDTO.getCourseMemo())
                 .courseDisclosure(saveDateCourseDTO.getCourseDisclosure())
                 .courseLike(saveDateCourseDTO.getCourseLike())
                 .courseView(saveDateCourseDTO.getCourseView())
@@ -48,7 +46,6 @@ public class DateCourseConverter {
         return DateCourse.builder()
                 .courseTitle(dateCourseDTO.getCourseTitle())
                 .courseType(dateCourseDTO.getCourseType())
-                .courseMemo(dateCourseDTO.getCourseMemo())
                 .courseDisclosure(dateCourseDTO.getCourseDisclosure())
                 .courseStartDate(dateCourseDTO.getCourseStartDate())
                 .courseEndDate(dateCourseDTO.getCourseEndDate())
@@ -61,7 +58,6 @@ public class DateCourseConverter {
                 .courseId(savedCourse.getCourseId())
                 .courseTitle(savedCourse.getCourseTitle())
                 .courseType(savedCourse.getCourseType())
-                .courseMemo(savedCourse.getCourseMemo())
                 .courseDisclosure(savedCourse.getCourseDisclosure())
                 .courseLike(savedCourse.getCourseLike())
                 .courseView(savedCourse.getCourseView())
@@ -79,7 +75,6 @@ public class DateCourseConverter {
                 .courseId(courseId)
                 .courseTitle(request.getCourseTitle())
                 .courseType(request.getCourseType())
-                .courseMemo(request.getCourseMemo())
                 .courseDisclosure(request.getCourseDisclosure())
                 .courseStartDate(request.getCourseStartDate())
                 .courseEndDate(request.getCourseEndDate())
@@ -91,7 +86,6 @@ public class DateCourseConverter {
                 .courseId(updatedCourseDTO.getCourseId())
                 .courseTitle(updatedCourseDTO.getCourseTitle())
                 .courseType(updatedCourseDTO.getCourseType())
-                .courseMemo(updatedCourseDTO.getCourseMemo())
                 .courseDisclosure(updatedCourseDTO.getCourseDisclosure())
                 .courseStartDate(updatedCourseDTO.getCourseStartDate())
                 .courseEndDate(updatedCourseDTO.getCourseEndDate())
@@ -107,7 +101,6 @@ public class DateCourseConverter {
                         .courseId(course.getCourseId())
                         .courseTitle(course.getCourseTitle())
                         .courseType(course.getCourseType())
-                        .courseMemo(course.getCourseMemo())
                         .courseLike(course.getCourseLike())
                         .courseView(course.getCourseView())
                         .courseStartDate(course.getCourseStartDate())
@@ -122,7 +115,6 @@ public class DateCourseConverter {
                 .courseId(dateCourseDTO.getCourseId())
                 .courseTitle(dateCourseDTO.getCourseTitle())
                 .courseType(dateCourseDTO.getCourseType())
-                .courseMemo(dateCourseDTO.getCourseMemo())
                 .courseLike(dateCourseDTO.getCourseLike())
                 .courseView(dateCourseDTO.getCourseView())
                 .courseStatus(dateCourseDTO.getCourseStatus())
