+++ src/main/java/beyond/momentours/date_course/command/application/service/DateCourseCommandService.java
@@ -4,9 +4,8 @@ import beyond.momentours.date_course.command.application.dto.DateCourseDTO;
 import beyond.momentours.member.command.application.dto.CustomUserDetails;
 import jakarta.transaction.Transactional;
 
-import java.util.List;
-
-public interface DateCourseService {
+public interface DateCourseCommandService {
+    @Transactional
     DateCourseDTO createDateCourse(DateCourseDTO dateCourseDTO, CustomUserDetails user);
 
     @Transactional
@@ -14,13 +13,4 @@ public interface DateCourseService {
 
     @Transactional
     void deleteDateCourse(Long courseId, CustomUserDetails user);
-
-    @Transactional
-    List<DateCourseDTO> getDateCourses(String sortBy);
-
-    @Transactional
-    DateCourseDTO getDateCourse(Long courseId, CustomUserDetails user);
-
-    @Transactional
-    List<DateCourseDTO> getCoursesByMemberId(CustomUserDetails user);
 }
