+++ src/main/java/beyond/momentours/date_course/command/application/service/DateCourseCommandServiceImpl.java
@@ -12,20 +12,16 @@ import beyond.momentours.member.command.application.dto.CustomUserDetails;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
 import java.time.ZoneId;
-import java.util.List;
-import java.util.stream.Collectors;
 
 @Slf4j
 @Service("commandDateCourseService")
 @RequiredArgsConstructor
-public class DateCourseServiceImpl implements DateCourseService {
+public class DateCourseCommandServiceImpl implements DateCourseCommandService {
 
-    private final RedisTemplate<String, String> redisTemplate;
     private final DateCourseRepository dateCourseRepository;
     private final DateCourseConverter dateCourseConverter;
     private final DateCourseMapper dateCourseDAO;
@@ -78,41 +74,9 @@ public class DateCourseServiceImpl implements DateCourseService {
         log.info("데이트 코스 Soft Delete 완료: courseId={}, userId={}", courseId, user.getMemberId());
     }
 
-    @Transactional
-    @Override
-    public List<DateCourseDTO> getDateCourses(String sortBy) {
-        List<DateCourse> courses = dateCourseDAO.findCoursesWithSorting(sortBy);
-        return courses.stream()
-                .map(dateCourseConverter::fromEntityToDTO)
-                .collect(Collectors.toList());
-    }
-
-    @Transactional
-    @Override
-    public DateCourseDTO getDateCourse(Long courseId, CustomUserDetails user) {
-        String key = "course:view:" + courseId;
-        redisTemplate.opsForValue().increment(key, 1);
-
-        DateCourse dateCourse = dateCourseDAO.findActiveById(courseId).orElseThrow(() -> new CommonException(ErrorCode.NOT_FOUND_DATE_COURSE));
-        if (!dateCourse.getCourseDisclosure() && !dateCourse.getMemberId().equals(user.getMemberId())) throw new CommonException(ErrorCode.ACCESS_DENIED);
-
-        return dateCourseConverter.fromEntityToDTO(dateCourse);
-    }
-
-    @Transactional
-    @Override
-    public List<DateCourseDTO> getCoursesByMemberId(CustomUserDetails user) {
-        Long memberId = user.getMemberId();
-        List<DateCourse> courses = dateCourseDAO.findCoursesByMemberId(memberId);
-        return courses.stream()
-                .map(dateCourseConverter::fromEntityToDTO)
-                .collect(Collectors.toList());
-    }
-
     private void update(DateCourseDTO dateCourseDTO, DateCourse existingCourse, Long courseId) {
         existingCourse.updateCourseTitle(dateCourseDTO.getCourseTitle());
         existingCourse.updateCourseType(dateCourseDTO.getCourseType());
-        existingCourse.updateCourseMemo(dateCourseDTO.getCourseMemo());
         existingCourse.updateCourseDisclosure(dateCourseDTO.getCourseDisclosure());
         existingCourse.updateCourseStartDate(dateCourseDTO.getCourseStartDate());
         existingCourse.updateCourseEndDate(dateCourseDTO.getCourseEndDate());
