+++ src/main/java/beyond/momentours/date_course/command/domain/vo/response/ResponseCreateDateCourseVO.java
@@ -14,7 +14,6 @@ public class ResponseCreateDateCourseVO {
     private Long courseId;
     private String courseTitle;
     private CourseType courseType;
-    private String courseMemo;
     private Boolean courseDisclosure;
     private Long courseLike;
     private Long courseView;
