+++ src/main/java/beyond/momentours/date_course/command/domain/vo/response/ResponseDateCourseDetailVO.java
@@ -14,7 +14,6 @@ public class ResponseDateCourseDetailVO {
     private Long courseId;
     private String courseTitle;
     private CourseType courseType;
-    private String courseMemo;
     private Long courseLike;
     private Long courseView;
     private Boolean courseStatus;
